refactor: extract plugin-wide RequestIdentityService identity ladder
The who-is-this-request ladder (ClaimsPrincipal -> tag marker -> cookie ->
session-by-IP) moves out of SpoilerUserResolver into Services/Identity as a
feature-agnostic service returning candidates + a confidence tier, so any
feature can consume request identity through one documented choke point.
SpoilerUserResolver keeps only spoiler-specific state loading and delegates.
