+++ test-results/.last-run.json
@@ -1,6 +1,4 @@
 {
-  "status": "failed",
-  "failedTests": [
-    "54cc58e42297daaa24fb-50913d87378786128382"
-  ]
+  "status": "passed",
+  "failedTests": []
 }
